Search for aenea.json config in the NatLink user directory.
NatLink 4.1oscar introduced a new configuration item for storing NatLink files created by the user, called the UserDirectory setting. Prior to this change any custom files, such as Dragonfly grammars, needed to be stored in NatLink's MacroSystem directory. However, this directory is intended to be used for NatLink internals and mixing in user-specific files creates a mess. Using the newer UserDirectory setting is now the golden path for storing user-specific files.