Add Travis CI integration for the client unit tests.
Aenea's automated tests are, sadly, not great. We have a simple
integration test for the server that relies on X11, some basic unit
tests for the client library, and a bunch of untested code.

I don't think I'll be able to spend the time fully fixing this,
unfortunately, but we can at least set up CI for the tests we have to
prevent them from bit rotting again, and so that if other contributors
wish to add tests they can more easily do so.

I've also decided to start enforcing pyflakes clean for all new PRs
going forward; there is a list of exclusions in .nopyflakes for files
that would be impractical to fix for the time being. New PRs must not
introduce new pyflakes errors.

Both pyflakes and the client-side unit tests, such as they are, will now
run in Travis.